+++ scripts/scripts.js
@@ -1,3 +1,4 @@
+import { request as graphqlRequest } from 'graphql-request';
 import {
   sampleRUM,
   buildBlock,
@@ -335,3 +336,64 @@ export function onNavigate(pathName) {
           });
         }
       } */
+
+export function getItems(categoryId) {
+  const url = 'https://venia.magento.com/graphql'; // Replace with your GraphQL endpoint
+  const headers = {
+    // Add any custom headers you need, like authorization headers, etc.
+    Authority: 'venia.magento.com',
+    Store: 'wknd',
+    'Content-type': 'application/json',
+  };
+  const dataRaw = `
+  query{   
+    products(filter: { category_uid: { eq: ${categoryId}} }) {
+      items {
+        name
+        sku
+        url_key
+        is_returnable
+        image {
+          label
+          url
+        }
+        small_image{
+          label
+          url
+        }
+        swatch_image
+        price_range{
+        maximum_price{
+          final_price{
+            currency
+            value
+          }
+        }
+        minimum_price{
+          final_price{
+            currency
+            value
+          }
+        }
+        }
+      }
+        page_info {
+          current_page
+          page_size
+          total_pages
+        }
+        total_count
+      }
+    }
+  `;
+  return graphqlRequest(url, dataRaw, null, headers)
+    .then((response) => {
+      if (!response) {
+        throw new Error('Network response was not ok');
+      }
+      return response;
+    })
+    .catch((error) => {
+      console.error('Error making the GraphQL request:', error.message);
+    });
+}
